modify: AppDelegate 구분, 오타 수정
