+++ hometerview/App/hometerviewApp.swift
@@ -7,13 +7,11 @@
 
 import SwiftUI
 
-import KakaoSDKCommon
 import KakaoSDKAuth
-import KakaoSDKUser
 
 @main
 struct hometerviewApp: App {
-    @UIApplicationDelegateAdaptor var dleegate: AppDelegate
+    @UIApplicationDelegateAdaptor var delegate: AppDelegate
     @StateObject private var user = User.shared
 
     var body: some Scene {
@@ -34,22 +32,5 @@ struct hometerviewApp: App {
                 }
             }
         }
-    }
-}
-
-class AppDelegate: NSObject, UIApplicationDelegate {
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
-
-        KakaoSDK.initSDK(appKey: "f811c290471ec82c6e65a3c0c68a3bbc")
-
-        return true
-    }
-
-    func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
-        if (AuthApi.isKakaoTalkLoginUrl(url)) {
-            return AuthController.handleOpenUrl(url: url)
-        }
-
-        return false
     }
 }
